requirements.txt
# Core ML and Computer Vision
tensorflow>=2.8.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.5
numpy>=1.21.0
Pillow>=8.3.0

#Newly installed frameworks
tf-keras
deepface
facenet-pytorch
dlib
scipy
librosa
sounddevice
soundfile
pyaudio
transformers


# GUI Framework
PyQt5>=5.15.6
qt-material>=2.12

# Object Detection
ultralytics>=8.0.0  # For YOLO
scikit-learn>=1.0.0

# Image Processing
matplotlib>=3.4.3
albumentations>=1.1.0

# Video Processing
moviepy>=1.0.3
imageio>=2.19.0

# Utility
pandas>=1.3.0
tqdm>=4.62.0
requests>=2.26.0